# api.yaml
openapi: 3.0.3
...
info:
  title: Placement Portal Application API
  description: >
    REST/JSON API for the Placement Portal (Flask + SQLite + Redis + Celery).
    All endpoints are under /api. Authentication uses a JWT bearer token
    obtained from /api/login and sent as "Authorization: Bearer <token>".
    Validation failures return 422 with {"errors": {field: message}}.
  version: 2.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: 'Validation failed — body is {"errors": {field: message}}'
    Unauthorized:
      description: Missing or invalid JWT
    Forbidden:
      description: Authenticated but wrong role / not permitted
security:
  - bearerAuth: []
paths:
...
  /api/register/student:
    post:
      tags: [Auth]
      security: []
      summary: Register a new student account
      description: Creates a User (role=student) and StudentProfile. Fields — email, password, name, branch, cgpa, graduation_year, phone (optional).
      responses:
        "201": { description: Registration successful }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/register/company:
    post:
      tags: [Auth]
      security: []
      summary: Register a new company (awaits admin approval)
      description: Creates a User (role=company) and CompanyProfile with approved=false. Company cannot log in until the admin approves it.
      responses:
        "201": { description: "Registration submitted, pending admin approval" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/login:
    post:
      tags: [Auth]
      security: []
      summary: Authenticate user and issue a JWT token
      description: Verifies email/password; rejects blacklisted, deactivated, or unapproved-company accounts. Returns {token, user}.
      responses:
        "200": { description: JWT token and user object }
        "401": { description: Invalid email or password }
        "403": { description: Blacklisted / deactivated / company pending approval }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user
      responses:
        "200": { description: Current user object }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # -------------------------------------------------------------- Common ----
...
  /api/company/dashboard:
    get:
      tags: [Company]
      summary: Company profile and its drives
      responses:
        "200": { description: "{profile, drives[]} with applicant counts" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/company/drive:
    post:
      tags: [Company]
      summary: Create a new drive (status Pending)
      description: Requires the company to be approved. Fields — title, description, eligible_branches[], min_cgpa, eligible_years[], deadline. Needs admin approval before students see it.
      responses:
        "201": { description: Drive created (Pending) }
        "403": { description: Company not yet approved }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/company/drive/{drive_id}:
    put:
      tags: [Company]
      summary: Update a drive (returns to Pending for re-approval)
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: "Drive updated, status reset to Pending" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/company/drive/{drive_id}/close:
    post:
      tags: [Company]
      summary: Close a drive to further applications
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Drive closed }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/company/drive/{drive_id}/applicants:
    get:
      tags: [Company]
      summary: List applicants of a drive
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Applications for the drive }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/company/application/{application_id}:
    put:
      tags: [Company]
      summary: Update application status/interview/remarks
      description: Status follows Applied → Shortlisted → Selected/Rejected transitions. Setting or changing interview_datetime immediately emails the student (Celery).
      parameters:
        - { name: application_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Updated application }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/company/application/{application_id}/offer-letter:
    post:
      tags: [Company]
      summary: Generate offer letter PDF (Selected only)
      parameters:
        - { name: application_id, in: path, required: true, schema: { type: integer } }
      responses:
        "201": { description: Offer letter generated }
        "200": { description: Offer letter already existed }
        "422": { description: Application is not in Selected status }
  /api/company/application/{application_id}/offer-letter/email:
    post:
      tags: [Company]
      summary: Email the offer letter to the student
      description: Queues a Celery task that emails the generated PDF to the student as an attachment.
      parameters:
        - { name: application_id, in: path, required: true, schema: { type: integer } }
      responses:
        "202": { description: Email queued for sending }
        "422": { description: Offer letter not generated yet }

  # ------------------------------------------------------------- Student ----
...